Re-implement the preservation of scroll state on room switch
This now happens slightly differently - we use the 'initialEventId' property
introduced in PR #96 (as well as a new 'initialEventPixelOffset' property) to
initialise the scroll state when returning to a room.

This commit also ensures that we scroll to the 'read marker' when switching to
a room which has no previous scroll state, as well as preventing that marker
from going past the middle of the screen.