Use the same deviceId after login
Make sure that we use the same deviceId *after* we have logged in as we used to
do the login - otherwise we end up with two devices registered on the server.

So, don't regenerate a deviceId on calls to
MatrixClientPeg.replaceUsingAccessToken, and instead pass it in from the old client.