Unpaginate gratuitously after paginating
This is in order to unpaginate based on the clientHeight of the events rendered, instead of assuming that each event takes up a fixed amount of vertical space, and that 250 events will be the maximum number of events to appear on the screen (within clientHeight).

This makes timeline-window effectively flexible, by preventing it from using its own unpagination by giving it a windowLimit of 10000.