Change method and architecture
As per @richvdh's suggestion, things have been moved around and changed. Unpagination is still done based on clientHeight of elements in the ScrollPanel, but now the ScrollPanel is the one calling for unpagination by it's property onUnfillRequest. This is used by the TimelinePanel to unpaginate events based on the first/last scrollToken that needs to be removed. ScrollTokens are event IDs and so the TimelinePanel simply assumes as much. It then takes a chunk of events off the front or back of this.state.events using timelineWindow.unpaginate.

The tests _are_ failling unfortunately, due to a scroll event not being emitted by the ScrollPanel on line 186. The if statement here, which is for a bug in Chrome on MacOS, prevents onScroll from being emitted, and the test cannot continue.