Hide menu onMouseLeave from RoomTile
This has the problem where the cursor might leave the tile before hitting the menu, hiding the menu. Also, when the cursor leaves the room avatar _and_ the room tile at the same time, the target given in that event is the child, the avatar, which cannot be used to hide the menu.