Maintain file order when uploading files.
This may address vector-im/riot-web#1459

Upload order is acheived by using a collection of promises that are used to block the promise chain. The chain is only blocked until the upload immediately prior to the desired upload has been completed. Blocking the promise chain is to still permit concurrent uploads, otherwise the sending of files would have been blocked instead.

Signed-off-by: Travis Ralston <travpc@gmail.com>