Expose isRMSet instead of conflating state
This exposes enough state from TimelinePanel to determine whether an RM has ever been set in a room (which implies that it has been joined). This seems monumentally better than trying to rewrangle getRMPos to fake the RM being on screen.

Also, I noticed the RM could get trapped in a MELS somehow so I fixed that by allowing a MELS to have its own RM. I also noticed that in some circumstances the RM wasnt updated because `currentReadUpToEventIndex` didnt consider the RM event ID. This was also fixed.