Update header, create ReskindexPlugin
This is a webpack plugin that can be run as part of a webpack build process to generate component-index.js automatically without it being run manually.

The only caveat is that it will necessarily cause a second rebuild with the new index bundled, but this takes a small amount of time and is only required when a file is added/removed/renamed or if the bundle has only been built once.

The plugin uses names of the files that are bundled to check whether it needs to run again. This is how it determines whether files have been added/removed/renamed since the last reskindex.