[WIP] trust in chrome to do scroll anchoring
This does not exactly work because it fails when we're at the very top of the page and back paginating. It does however work very very well when paginating whilst the viewport is half-way down the timeline window or whilst at the very bottom.

Removing spinners doesn't seem to have any effect.

In theory the anchor selection algorithm chrome uses is selecting an element that React decides to turn into a totally new element. There is currently no planned mechanism to prevent some elements from being used as anchors.

Chome is also the only browser to support this

See https://drafts.csswg.org/css-scroll-anchoring/