Dismiss settings by clicking between panels
If the window was wide enough, clicking on the gap between the middle
panel and the side panels didn't work to dismiss settings as you'd
think it would.

This moves the wrapper elements to MatrixChat so they're always there
such even when there is no RightPanel (also means only one handler,
and both side panels don't need to worry about how they close the
settings. As a bonus, the middle panel now doesn't move around when
the right panel is minimised.