Show connection error message immediately, also during reconnect
As catchup sync can take quite some time (maybe temporarily even slower with LL)
it can be confusing when you come back to riot after a couple of days and you only see old messages while the first sync is running. If riot didn't get the time to fail /sync 3 times (like when sleeping your laptop), it won't show you anything to indicate your timeline is behind and its reconnecting.

This is probably not the best way to deal with this, as in the laptop wakeup scenario, you might as well have 0 failed syncs but still catchup takes minutes. Looking at how much time there was between the last and new sync might be better, but this sort of does the job for now