Ensure user directory lookups show the correctly merged results
This involves tracking each user we see (which will only be slow on the first execution) and ensuring we wait for that to complete. By waiting, we ensure that the getEffectiveParents() call can return the correct parent users for the results. Additionally, we need to use the right property off the results when preparing tuples for getEffectiveParents().

Because the user directory results should be relatively limited, we can get away with running all the tracking concurrently via the .map() function and awaiting on all the promises to complete.