Remove cancel button from final key backup screen
Canceling the dialog once the backup was created meant onFinished
was called with false which RoomView interprets as the backup not
being done.

This does mean you can't just hit esc to clear the dialog once you're
done: if we wanted this we could fix this in a better way.

https://github.com/vector-im/riot-web/issues/8066