Switch the guest account over to the right homeserver after discovery
In 5b2328bad37cb8780e2334602b59e62ee7c585e0 we log out of the guest account before moving on, however that doesn't solve the case of the guest account being wrong to begin with. We should still keep 5b2328bad37cb8780e2334602b59e62ee7c585e0 in the event this races, however this commit's goal is to use the right homeserver for the guest account once it has been discovered.