Treat SettingsDialog as variable width
In reality it sets its onw width as 90% or 1000px.

Another thing we could do here is change the fixedWidth prop to
BaseDialog to be a number which it would set the width to, then
we could remove the CSS overrides for SettingsDialog. Would
essentially be putting more style into the JS though.