Upgrade to babel@7 and support TypeScript
For https://github.com/vector-im/riot-web/issues/10524

Upgrading to babel 7 gets a lot of the compilation stuff for free. The major changes we care about are:
* The packages are now namespaced (hence all the package changes)
* We use the `env` preset instead of the long-since deprecated es20XX packages. 
* Added `--verbose` to maintain log output from the previous version.

This also meant fixing `ScalarAuthClient`: we can no longer mix `module.exports` and `import` in the same file. 

The SdkConfig was converted to TypeScript to prove this works (it does).