Set the noreferrer attribute to links and images
An alternative fix to vector-im/riot-web#6147 which for some reason
the PR vector-im/riot-web#6155 is not yet merged.

The key difference is that the riot-web PR vector-im/riot-web#6155
uses HTML meta header for noreferrer, while this one adds the
rel-attribute to include the noreferrer keyword in both user-created
links as well as links converted from incoming events.

I guess it's up to the maintainers then to pick and choose, but please
do ;).