A clearer GUI sequence and a basic verify password strength method are created.
To guide the user when requesting a new password, the submit button in invalidated until a (most likely) correct e-mail address,  a new password and password conformation are given. For each state a new description on the button is added. This blocks filling in usernames or matrixID or other strings which cannot be an e-mail address. It does not validate the existence of an e-mail address.
This should solve issue https://github.com/vector-im/riot-web/issues/9978