Fix jumping to read marker for events without tiles
It is possible for your read marker to be set to an event without a tile (like a
reaction). We would still render the read marker at the position of those events
in the timeline, even though there's no matching tile, which breaks a core
assumption of jump to read marker path: it assumes that if the read marker node
is present in the DOM, then there must also be an event tile with a scroll token
matching the event ID.

This fixes the situation by no longer attempting to draw a read marker for
events without tiles.

Fixes https://github.com/vector-im/riot-web/issues/10975