Fix stuttering when typing in font name
This fixes a major point of lag when a user changes the font used
in Element. When the "use system font" input is updated (by a
keystroke), the client updates the font, by dispatching a font
change event to the central dispatcher. One then assumes that
this changes the font family in the style sheet, as there is quite a
substantial amount of lag on every keystroke.

This commit fixes that problem, by introducing a "debounce" to the
"use system font" controller's dispatcher method. After some testing,
3000 milliseconds seemed to work the best, but that's easily changed.

Would it be a good idea to move the milliseconds outside of the
call to debounce?

Can't find any issues about this. That's surprising.

Signed-off-by: resynth1943 <resynth1943@tutanota.com>