Early structure for MSC1767
This contains the basic text parsing for [MSC1767](https://github.com/matrix-org/matrix-doc/pull/1767), though is not hooked up to anything yet. Tests and code are written such that they could be pulled out of the react-sdk if desirable - they are just here so it is faster to iterate on.