Allow ContextMenu instances to be simply mounted as children of the curr. container
By default, ContextMenu's are children of mx_ContextualMenu_Container, which is placed at the
root of the DOM. A new prop, mountAsChild, allows for ContextMenus to be mounted as children
of the current container instead.

This is necessary as mx_ContextualMenu_Container won't be included in the content that's
shown when a call is fullscreened, and could potentially be useful elsewhere.