Retry loading of the pickle key if the user is logged in.
If the Keyring is unavailable or not unlocked Element would start
without being able to decrypt messages. This fix will ensure Element
doesn't start unless it was able to retrieve the encryption key.

Unlocking the Keyring at a later time would not result in the messages
being decrypted. The app would need to be restarted to achieve this.

This fixes https://github.com/vector-im/element-web/issues/17845