Add more checkbox styles
Add a "kind" param for StyledCheckbox, allowing designers to choose
different styles of checkbox as needed.  In addition to the default
kind, there's now an "outline" kind with a green checkmark and a
transparent fill.  This is used in the device trust view, since the
default checkbox style looks too much like the green "verified" shield
and it's awkward to have those next to each other.