Restore 'large' thumbnails to be 800x600
In order to restore the original behaviour for those who want it, redefine "large" thumbnails to be fit within 800x600px.
This means that timeline screenshots end up typically not being downscaled (which is important given screenshots are our easiest way of copying chunks of history between rooms currently :/).  It also means that photos end up nice and big, which was a deliberate aesthetic choice for the app in the past, which some users will want to retain.