Force ContextMenu to re-render regardless if any rooms moved around from the sort change
See https://github.com/matrix-org/matrix-react-sdk/pull/7353#discussion_r769995470

Previously, the ContextMenu would only update and render if shouldComponentUpdate
said it should. This excludes the options mutated by the ContextMenu. We forceUpdate
for the other things in the ContextMenu.

Previously, the ContextMenu only rendered properly when it also changed the sort
of the rooms in the list which triggered a render as well.