Add labs flag for default open right panel
This adds an experimental default open right panel mode as a quick fix for those
who prefer to have the right panel open consistently across rooms.

If no right panel state is known for the room or it was closed on the last room
visit, it will default to the room member list. Otherwise, the saved card last
used in that room is shown.

Fixes https://github.com/vector-im/element-web/issues/20666