Add body link fallback handler for in-app navigation
This adds a fallback click handler for all links in the message to ensure that
links that were ignored as part of linkification (e.g. pills, links in message
content) will still navigate in-app where possible.

This approach preserves a side benefit of
https://github.com/matrix-org/matrix-react-sdk/pull/7453, which kept all link
hrefs as matrix.to style links (so that they can be easily copied and shared).

Fixes https://github.com/vector-im/element-web/issues/20715