restructure legacy/custom scss files
to make sure scss variables are being properly loaded and propagated the import order has been corrected. it is now changed to [common hardcoded values (_legacy-default) -> theme specific hardcoded values (_legacy-dark/light) -> custom variable values (_legacy-custom) -> common dependant values (_legacy-common)]. in order to facilitate this import order values that were hardcoded in _legacy-dark but not in _legacy-light are now hardcoded in _legacy_light and vice versa. variables were changed to be the exact value of the hex so no change in colors is in this commit. it is important to note that this code does slightly change the custom theming api. in particular it removes exposed variables in the api that were being rewritten (i.e. bugged variables) users of the api should not see anything change in theming only that some variables (that had no effect anyway) cannot be set".