Move LeftPanelLiveShareWarning+functionality into LoggedInView
Annoyingly this increases complexity of an already-complex LoggedInView, however the flex layout in the left panel doesn't allow for us to easily add a rowspan-like element such as the live location share warning. 

To circumvent this for purposes of release, we'll just chuck the live location sharing warning functionality into LoggedInView, which can manage a set of CSS classes to mimic a rowspan across the top of the room list.

Trying to make the left panel a nested flexbox works visually, but breaks all the scroll behaviour for the room list and space panel. This is likely because the flexbox layout of the left panel is already trying to do too many tricks to get the layout to work just right, so adding a fifteenth trick by nesting the flexboxes just doesn't work. It's unclear which trick specifically is breaking the scroll behaviour.

It's absolutely awful, but it works.