Replace padding with margin for an image and a file button for group layout on ThreadView
This normalizes inline spacing of image and file button on ThreadView.

The declarations of EventTile_mediaLine was added with 3a2aa9d, and they were expected to align the element with mx_EventTile_content. After the padding-right value for mx_EventTile_content was changed from 50px to 8px with 544957b, the declarations have been obsolete.

Signed-off-by: Suguru Hirahara <luixxiul@users.noreply.github.com>