settings: use composition for VoiceUserSettingsTab's state
IState does not seem to be in an "IS-A" relation with
MediaDeviceKindEnum-based records (Liskov substitution principle), so
applying the "composition over inheritance" concept might be a good idea.

This can be refactored further by adding separate types for I/O device
options and for audio settings and then using them in IState via
composition.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>