Add a larger buffer to useLatestResult's test
Should fix https://github.com/vector-im/element-web/issues/23014

With only a 5ms buffer previously it was possible for Jest to get overwhelmed with the 1500 tests it needs to run. We raise this to 20ms to give extra time to settle.

Even with the high sleep times this won't be the longest running test, so should be fine.