Fix scroll-jumps when jumping to a permalink
Having turned on scroll debugging, it looks like
https://github.com/vector-im/element-web/issues/23393 is caused
by scroll events firing while forward-filling is happening, which
causes ScrollPanel's onScroll() to lock the scroll position to the
bottom of the page via saveScrollState(), rather than honour the
permalink's scroll offset.

Either we could make saveScrollState() honour permalinks, or
just ignore scroll events which fire while filling - as checkFillState()
will go and apply the scroll-lock anyway once the filling is complete.

This fix attempts the latter; i've manually tested a bunch (~20 times)
and it /seems/ to have fixed it? It was pretty hard to repro in the first
place, though.