Add composer option to autoformat selection as list
This adds a button and a shortcut for the ability to format selected lines as either
a bulleted or a numbered list in the composer. Similar to other common implementations
of this feature, it also allows to format complete lines if they are only partially
selected, or single lines at caret position without any selection. Pressing the button
or shortcut again or while selecting already formatted sections will unformat the lines.