Don't wrap list elements in paragraph tags
Under certain circumstances (see issue linked below), the content of list elements are wrapped in paragraph tags, resulting in unintentional styling. This PR provides a workaround by not allowing nodes in a "list" to get wrapped in paragraph tags.

To be honest, I doubt this is the optimal solution. But it is at least a solution that can hopefully prove at least as a starting point on how to resolve the accompanying issue.

Resolves: https://github.com/vector-im/element-web/issues/23507

Signed-off-by: Arne Wilken arnepokemon@yahoo.de