make the message composer restricted by vh not px
The size of the composer if very limiting when writing a longer
message with a lot of details and links.

This change let's it grow up until 70% of the current view height
so the autocomplete is still visible, but we can also see as much
of the message as possible.