Add a URL panel to the right panel
Adds a new panel the right menu that, akin to "Files", shows all
messages in the room containing at least one URL. This would
help users find previously posted URLs without having to come
up with the correct search term(s) to find them.

Code-wise this more of a proof of concept. `UrlPanel.tsx` is almost
completely copy and pasted from `FilePanel.tsx`. `UrlPanel.tsx`
should probably either be stripped down or both of them should
be abstracted into a generic class.

Furthermore, the type of `filter` necessary for this PR does not seem to
exist (the one suggested by this PR is probably quite a hack), and I am
unsure what the best course of action in creating such a filter would be.
Any advice on that and on how/why synapse should be involved would
be appreciated.

Signed-off-by: Arne Wilken arnepokemon@yahoo.de