Allow IndexedDB cleanup operations to fail
Firefox ships with a preference (`dom.indexedDB.privateBrowsing.enabled`) which
causes `window.indexedDB` to fail on all requests including `deleteDatabase`.
Because of that, Firefox Private Browsing users are left with an infinite
loading spinner.

`MatrixClient#clearStores` will resolve to a rejected promise if any of its
operations fails, which seems correct -- it's unable to complete its tasks.
So, to keep from erroring out during initialization, the errors have to be
discarded during the lifecycle operations.

Signed-off-by: Clark Fischer <clark.fischer@gmail.com>