Move subscriptions out of RoomView constructor
Components are not 'cleaned up' (i.e. `componentWillUnmount` called) unless
`componentDidMount` was called. React doesn't expect to have to simulate
an unmount if the component was never mounted to begin with.

In the case of RoomView, that means that the subscriptions created in the
constructor are never removed, i.e. a leak.

This moves the subscriptions to the `componentDidMount` method.

See https://github.com/vector-im/element-web/issues/24236#issuecomment-1397428515

Signed-off-by: Clark Fischer <clark.fischer@gmail.com>