Rename: apply our naming policy - `mx_InviteDialog`
There are two variants of invite dialogs; --transfer, --other, --hasFooter.

The dialog for transfer is related to the dialpad. The other dialog is one for inviting to a room and a DM. The dialog with footer is one with a link to your Matrix ID. They should be named as variants and styled accordingly.

- mx_InviteDialog
- mx_InviteDialog--transfer
- mx_InviteDialog--other
- mx_InviteDialog--hasFooter