Remove redundant selectors
Because "mx_AppTile" class name is now globally applied, it has become redundant to apply "mx_AppTileBody" and "mx_AppTileBody--mini" to "mx_AppTile--fullWidth" and "mx_AppTile--mini", respectively. It is now fine to apply them to "mx_AppTile".

Another theoretical possibility is a combination of "mx_AppTile--fullWidth" AND "mx_AppTileBody--mini", but it does not occur in reality because "mx_AppTile--fullWidth" is not applied on miniMode.